💄 Add custom button for load Excel file
