+++ src/components/ReadExcelFile.tsx
@@ -1,9 +1,9 @@
+import { Group } from "@mantine/core";
 import React, { useState } from "react";
-import { read, utils } from "xlsx";
 import { IProject } from "../utils/interface/IProject";
-import { ProjectParameters } from "../utils/ProjectParameters";
 import ExcelDataGrid from "./dataGrid/ExcelDataGrid";
 import ExcelDataGridSettings from "./dataGrid/ExcelDataGridSettings";
+import LoadExcelFileButton from "./LoadExcelFileButton";
 
 const ReadExcelFile = () => {
   const [project, setProject] = useState<IProject[]>();
@@ -14,61 +14,15 @@ const ReadExcelFile = () => {
     "RESSOURCES",
   ]);
 
-  const getData = async (e: React.ChangeEvent<HTMLInputElement>) => {
-    const projects = [];
-    if (e.currentTarget.files === null) {
-      return console.log("Fichier manquant...");
-    } else {
-      const blob = new Blob([e.currentTarget.files[0]]);
-      const file = await blob.arrayBuffer();
-      const wb = read(file);
-      const data = utils.sheet_to_json<IProject>(wb.Sheets[wb.SheetNames[0]]);
-
-      for (let i = 0; i < data.length; i++) {
-        const element = data[i];
-
-        for (let j = 0; j < ProjectParameters.length; j++) {
-          if (element[ProjectParameters[j]] === undefined) {
-            element[ProjectParameters[j]] = "";
-          }
-
-          if (ProjectParameters[j] === "LIVR. (tri)") {
-            element.LIVRTRI = element["LIVR. (tri)"];
-          }
-        }
-        const agenceLetter = element.DOSSIER.slice(-1);
-        switch (agenceLetter) {
-          case "L":
-            element.AGENCE = "Clisson (44)";
-            break;
-          case "U":
-            element.AGENCE = "Anglet (64)";
-            break;
-          case "Y":
-            element.AGENCE = "Villefranche-sur-Saône (69)";
-            break;
-
-          default:
-            element.AGENCE = "Autre";
-            break;
-        }
-        projects.push(element);
-      }
-      e.target.value = "";
-      setProject(projects);
-    }
-  };
   return (
     <>
-      <input
-        type="file"
-        onChange={getData}
-        accept=".csv, application/vnd.openxmlformats-officedocument.spreadsheetml.sheet, application/vnd.ms-excel"
-      />
-      <ExcelDataGridSettings
-        showParams={showParams}
-        setShowParams={setShowParams}
-      />
+      <Group>
+        <LoadExcelFileButton project={project} setProject={setProject} />
+        <ExcelDataGridSettings
+          showParams={showParams}
+          setShowParams={setShowParams}
+        />
+      </Group>
       <h2>{project?.length !== undefined ? project?.length : 0} Affaire(s)</h2>
       <ExcelDataGrid project={project} showParams={showParams} />
     </>
